+++ src/main/java/com/hsu/shimpyoo/domain/chatbot/service/ChatServiceImpl.java
@@ -32,28 +32,6 @@ public class ChatServiceImpl implements ChatService {
     @Value("${gpt.api.key}")
     private String apiKey;
 
-    public ResponseEntity<CustomAPIResponse<?>> makeChatRoom(){
-        // 현재 인증된 사용자의 로그인 아이디를 가져옴 (getName은 loginId를 가져오는 것)
-        String loginId= SecurityContextHolder.getContext().getAuthentication().getName();
-
-        // 사용자 존재 여부 확인
-        Optional<User> isExistUser=userRepository.findByLoginId(loginId);
-        if(isExistUser.isEmpty()){
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"존재하지 않는 사용자입니다.");
-        }
-
-        ChatRoom chatRoom = ChatRoom.
-                builder()
-                .chatTitle("채팅방")
-                .userId(isExistUser.get())
-                .build();
-
-        chattingRoomRepository.save(chatRoom);
-
-        CustomAPIResponse<Object> res=CustomAPIResponse.createSuccess(200,  null, "채팅방이 생성되었습니다.");
-        return ResponseEntity.status(HttpStatus.OK).body(res);
-    }
-
 
     public ResponseEntity<CustomAPIResponse<?>> askForChat(String question) {
         try {
@@ -100,35 +78,5 @@ public class ChatServiceImpl implements ChatService {
             e.printStackTrace();
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR,"API 호출 중 오류 발생 : " + e.getMessage());
         }
-    }
-
-    @Override
-    public ResponseEntity<CustomAPIResponse<?>> modifyChatRoomTitle(ModifyChatRoomTitleDto requestDto) {
-        // 현재 인증된 사용자의 로그인 아이디를 가져옴 (getName은 loginId를 가져오는 것)
-        String loginId= SecurityContextHolder.getContext().getAuthentication().getName();
-
-        // 사용자 존재 여부 확인
-        Optional<User> isExistUser=userRepository.findByLoginId(loginId);
-        if(isExistUser.isEmpty()){
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"존재하지 않는 사용자입니다.");
-        }
-
-        Optional<ChatRoom> isExistChatRoom = chattingRoomRepository.findById(requestDto.getChatRoomId());
-
-        if(isExistChatRoom.isEmpty()){
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"존재하지 않는 채팅방입니다.");
-        }
-
-        // 채팅방이 현재 로그인한 사용자의 채팅방이 아니라면
-        if(isExistChatRoom.get().getUserId()!=isExistUser.get()){
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN,"채팅방 제목 수정 권한이 없습니다.");
-        }
-
-        isExistChatRoom.get().setChatTitle(requestDto.getTitle());
-        chattingRoomRepository.save(isExistChatRoom.get());
-
-        CustomAPIResponse<Object> res = CustomAPIResponse.createSuccess(200, null ,
-                "채팅방 제목이 수정되었습니다.");
-        return ResponseEntity.status(HttpStatus.OK).body(res);
     }
 }
